requirements.txt
requests
SQLAlchemy
PyMySQL
python-dotenv 